requirements.txt
fastapi
uvicorn
tensorflow
firebase-admin
pillow